Properly forward stdout, stderr and exit status in the launchers
The COMPILE_LAUNCHER and LINK_LAUNCHER scripts discarded the output
of the command they ran, which caused us not to have any useful error
output in some cases. I don't actually understand why there was some
output at all, but at least now we're doing things properly.

Also, we don't need to pre-process the arguments anymore, since we
launch the command directly instead of using a subshell. I find this
to be a superior alternative.