Transpose measurements in Javascript instead of Ruby
This moves all data massaging down into the Javascript, and uses Ruby only
for data measurement. This makes the code easier to follow, because the Ruby
code only performs measurements and gives that as-is to the Javascript.
Before, we had to keep track of which stage of the script used which data
representation, which was confusing.